refactor(upload): extract audit helpers to uploadAudit.ts; dedupe confirmation payload type
- route.ts passe sous 400 lignes (helpers d'audit déplacés dans uploadAudit.ts)
- SecondDeclarationConfirmationPayload = alias de DeclarationConfirmationPayload (DRY)

Refs #3782
